Jpa Delete Beer By Id Not Found

+++ src/main/java/guru/springframework/spring6restmvc/services/BeerServiceJPA.java
@@ -57,8 +57,12 @@ public class BeerServiceJPA implements BeerService {
     }
 
     @Override
-    public void deteteById(UUID beerId) {
-        beerRepository.deleteById(beerId);
+    public Boolean deleteById(UUID beerId) {
+        if (beerRepository.existsById(beerId)) {
+            beerRepository.deleteById(beerId);
+            return true;
+        }
+        return false;
     }
 
     @Override
